requirements.txt
Flask==2.3.0
Flask-MySQLdb==1.0.1
mysqlclient==2.1.1
Werkzeug==2.3.0
reportlab==4.0.4
openpyxl==3.1.2
twilio==8.5.0
Pillow==10.0.0
python-dotenv==1.0.0
gunicorn==21.2.0